Misc stress testing components
bigfile-generator spins up a huge dummy input file with 100 components and 100 vulnerabilities, with each component having 1 to 3 vulnerabilities.  bf1, bf2 and bf3 are outputs from this script.  They are then fed into the FTG to test how it handles large input using the shell script, to generate the files output1, output2 and oputput3.

(Note that the weights in the final output files' cutsets are just the sum of the cutset vulnerability weights for the purposes of this test run, because I sent them to a fellow student who's project program was expecting integer weights.)